Move eigen decomp into eigen module
